Modify redeployment script to deploy portfolio as a service and improve error handling.

+++ redeploy-site.sh
@@ -1,20 +1,36 @@
 #!/bin/bash
 
-SESSION_NAME="portfolio_server"
+cd ~
+if [ ! -d portfolio ]; then
+	echo "Could not find directory '~/portfolio'"
+	echo "Cloning portfolio from github..."
+	git clone https://github.com/DavidBenj15/portfolio.git portfolio
+fi
 
-cd ~/portfolio
+echo "Changing into portfolio directory"
+cd portfolio
 
-# Kill all existing tmux sessions
-tmux kill-server
-
-# Fetch latest changes from GitHub
+echo "Fetching latest changes from GitHub..."
 git fetch && git reset origin/main --hard
 
-# Activate venv and install dependencies
+if [ ! -f ".env" ]; then
+	pwd
+	ls -la
+	echo "Error: .env file not found. Exiting."
+	exit 1
+fi
+
+if [ ! -d python3-virtualenv ]; then
+	echo "Could not find directory 'python3-virtualenv'"
+	echo "Creating new virtual environment 'python3-virtualenv'..."
+	python -m venv python3-virtualenv
+fi
+
+echo "Activating virtual environment and installing dependencies..."
 source python3-virtualenv/bin/activate
 pip3 install -r requirements.txt
 
-# Start server on detatched tmux session
-tmux new-session -d -s $SESSION_NAME 'cd ~/portfolio && source python3-virtualenv/bin/activate && cd app && export FLASK_APP=__init__.py && flask run --host=0.0.0.0 > ../flask.log 2>&1'
+echo "Restarting 'myportfolio.service' service..."
+systemctl restart myportfolio.service
 
-
+echo "Success! Portfolio has been successfully redeployed."
